Status after 1st prompt
Week 7, task 5

Prompt:
I was to create a web user interface for the rock, paper, scissors application.
This is a Python project built using Poetry. Reuse as much of the current code
as possible. Player number one is always a human player. Player two can be a
human player or an AI.